head,big.matrix-method {bigmemory}	R Documentation

Return First or Last Part of a big.matrix Object

Description

Returns the first or last parts of a big.matrix object.

Usage

## S4 method for signature 'big.matrix'
head(x, n = 6)

## S4 method for signature 'big.matrix'
tail(x, n = 6)

Arguments

x
A big.matrix object

n
A single integer for the number of rows to return

[Package bigmemory version 4.6.4 Index]