print,big.matrix-method {bigmemory}	R Documentation

Print Values

Description

print will print out the elements within a big.matrix object.

Usage

## S4 method for signature 'big.matrix'
print(x)

Arguments

x
A big.matrix object

Note

By default, this will only return the head of a big.matrix to prevent console overflow. If you turn off the bigmemory.print.warning option then it will convert to a base R matrix and print all elements.

[Package bigmemory version 4.6.4 Index]